requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-FontAwesome==0.1.2
Flask-SQLAlchemy==2.4.0
Flask-Turbolinks==0.2.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
pyOpenSSL==19.0.0
python-slugify==3.0.2
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.3
text-unidecode==1.2
urllib3==1.25.2
uWSGI==2.0.18
Werkzeug==0.15.2
WTForms==2.2.1
